implementing creation of a product and refresh token fix

+++ gulait-mongo/api/routes/logout.js
@@ -10,7 +10,8 @@ const joi = require( 'joi' );
  */
 router.delete( '/', authenticateRefreshToken, async ( req, res ) => {
     const validationSchema = new joi.object().keys( {
-        userName: joi.string().required()
+        userName: joi.string().required(),
+        roles: joi.array()
     } );
 
     joi.validate( req.body, validationSchema, ( err, results ) => {
@@ -20,19 +21,21 @@ router.delete( '/', authenticateRefreshToken, async ( req, res ) => {
         } );
     } );
 
+    //if response has already been sent, 
+    //dont process any further
+    if( !res.headersSent ){
+        try {
+            const returnedRefreshToken = await RefreshToken.find( { userName: req.body.userName } );
+            console.log( `logout here userName = ${ req.body.userName }, returned RefreshToken = ${ returnedRefreshToken }` );
+            if( returnedRefreshToken.length === 0 ) return res.status( 403 ).json( { message: 'ERROR', data: "Refresh token doesn't exist" } );
+            const removedRefreshToken = await RefreshToken.deleteOne( { userName: req.body.userName } );
 
-    try {
-        const returnedRefreshToken = await RefreshToken.find( { userName: req.body.userName } );
-        console.log( `logout here userName = ${ req.body.userName }` );
-        if( returnedRefreshToken.length === 0 ) return res.status( 403 ).json( { message: 'ERROR', data: "Refresh token doesn't exist" } );
-        const removedRefreshToken = await RefreshToken.deleteOne( { userName: req.body.userName } );
-
-        res.status( 200 ).json( { message: 'SUCCESS', data: `Successfully logged out ${ req.body.userName }` } );
-    } catch (err) {
-        console.log( err );
-        res.status( 500 ).json( { message: 'ERROR', data: err.toString() } );
+            res.status( 200 ).json( { message: 'SUCCESS', data: `Successfully logged out ${ req.body.userName }` } );
+        } catch (err) {
+            console.log( err );
+            res.status( 500 ).json( { message: 'ERROR', data: err.toString() } );
+        }
     }
-        
 } )
 
 module.exports = router;
